requirements.txt
psycopg2-binary
numpy==1.21.6